requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
psycopg2==2.9.9
tqdm==4.66.2